Rename: 디렉토리 내 파일명 변경 및 이동
- PascalCase > kebab-case 로 변경
- 공통 컴포넌트와 분리하고 해당 디렉토리 아래로 이동하여 구조화
