--- src/app/contact/ContactContent.tsx
@@ -1,3 +1,0 @@
-export default function ContactContent() {
-  return <div>이메일 보내는 부분</div>;
-}
