+++ src/app/contact/contact-content.tsx
@@ -1,0 +1,10 @@
+import ContactForm from "./contact-form";
+
+export default function ContactContent() {
+  return (
+    <div>
+      <h3>이메일 보내기</h3>
+      <ContactForm />
+    </div>
+  );
+}
