+++ src/app/contact/page.tsx
@@ -1,5 +1,5 @@
 import CommonFrame from "@/components/CommonFrame";
-import ContactContent from "./ContactContent";
+import ContactContent from "./contact-content";
 
 export default function Contact() {
   return <CommonFrame sectionName="Contact" content={<ContactContent />} />;
