+++ src/app/feedback/page.tsx
@@ -1,5 +1,5 @@
 import CommonFrame from "@/components/CommonFrame";
-import FeedbackContent from "./FeedbackContent";
+import FeedbackContent from "./feedback-content";
 
 export default function Feedback() {
   return <CommonFrame sectionName="Feedback" content={<FeedbackContent />} />;
